% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db.R
\name{check_db}
\alias{check_db}
\title{Check neuroimaGene database downloaded}
\usage{
check_db(timeout = 900)
}
\arguments{
\item{timeout}{time to spend downloading the NeuroimaGene database in seconds (default = 900)}
}
\value{
no return value, called to give information on status of neuroimaGene
database and prompt user the user to download if resource file is missing.
}
\description{
Check if the NeuroimaGene database exists in the proper location prior to running the query and prompt user to download if not.
}
\examples{
check_db(timeout = 600)
}
\keyword{installation}
