% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_vignette_doc.R
\docType{data}
\name{ng_vignette}
\alias{ng_vignette}
\title{NeuroimaGene object for vignette illustration}
\format{
A data.table with 6 columns and 3824 rows:
\describe{
  \item{gene}{character: ENSEMBL Gene ID}
  \item{gene_name}{character: HUGO gene name}
  \item{gwas_phenotype}{character: neuroimaging derived phenotype}
  \item{training_model}{character: JTI derived tissue gene expression model}
  \item{zscore}{numeric: normalized effect size of GReX on NIDP morphology}
  \item{mod_BHpval}{character: Benjamini Hochberg corrected pvalue corrected by modality }
}
}
\source{
Bledsoe, X. (2024) A transcriptomic atlas of the human brain reveals genetically determined aspects of neuropsychiatric health
}
\usage{
data(ng_vignette)
}
\value{
This script has no return. This is a documentation file for the
neuroimaGene data subset required to build the vignette.
}
\description{
NeuroimaGene data table containing select associations used in the package vignette.
}
